requirements.txt
aiogram==2.25.2
openai==0.28.1
python-dotenv
aiohttp==3.8.6
